+++ src/index.html
@@ -103,8 +103,10 @@
       </div>
     </div>
     </main>
-    <div id="cards"></div>
-    <div id="root"></div>
+    
+    <div id="root">
+      <div id="cards"></div>
+    </div>
     <script src="main.js" type="module"></script>
   </body>
 </html>
